Fixing SSL errors during `requests` transaction in Pyinstaller context
`pyi_ssl_certs_workaround` will do something only in a Pyinstaller context on Linux. It will
search for a valid **SSL_CERT_DIR** or a **SSL_CERT_FILE** and replace process image
to allow openssl to re-init itself with correct x509 cert chain.

Everything is transparent for the user.

For more info on *ssl_files* and *ssl_dirs* vars, see:
https://github.com/golang/go/blob/master/src/crypto/x509/root_linux.go